+++ llvm/docs/ReleaseNotes.rst
@@ -98,6 +98,8 @@ Changes to the PowerPC Backend
 Changes to the RISC-V Backend
 -----------------------------
 
+* Zihintntl extension version was upgraded to 1.0 and is no longer experimental.
+
 Changes to the WebAssembly Backend
 ----------------------------------
 
@@ -166,9 +168,6 @@ Changes to the LLVM tools
 
 Changes to LLDB
 ---------------------------------
-
-* AArch64 Linux targets now provide access to the Thread Local Storage
-  register ``tpidr``.
 
 Changes to Sanitizers
 ---------------------